APM: add trace-agent.benchmarks to invoke scripts, fix frozen BenchmarkAgentTraceProcessing (#14465)
We need trace-agent.benchmarks available on the `main` branch so we can run comparisons as we
work on adding automated comparative benchmarks. Also the BenchmarkAgentTraceProcessing has been
broken for quite some time it seems as it needs the test agent to be `Run` so stats payloads are
retrieved from the channel.