pkg/{otlp,trace}: introduce APM stats consumer (#14586)
This change introduces an APM Stats consumer into the OTLP Ingest
pipeline to ensure that APM Stats coming is as OTLP Metrics generated by
the Datadog Processor in the OpenTelemetry Collector correctly get
flushed.

The logic for converting between APM Stats and Metrics that is used by
both the OTLP Ingest and datadogprocessor lives in the shared
pkg/otlp/model package.