pkg/ofbfuscate: allow dollar sign in SQL identifiers (#14639)
The SQL Obfuscator has not allowed dollar signs as part of identifiers.
However, dollar signs are valid identifier characters in some SQL
variants. This has resulted in valid queries being marked unparsable.

This commit includes the dollar sign in the set of allowed identifier
characters when tokenizing a SQL query.