(RCM-539) Add RC service and client to cluster-agent (#14276)
The "agent client" implementation was hard-coded to only work over a
secure gRPC client provided by a utility package in the datadog core
agent. We have a use case coming up where we want to be able to run both
the core remote config service AND an agent client in the same process
without having to setup and run the secure gRPC server within that
process.

In order to do this, we define an interface for the main config update
request and pull out the gRPC logic into a type that implements this
interface. We provide a convenience constructor to create an RC agent
client backed by the secure gRPC client to prevent the gRPC constructor
from needing to leave the internal package.

With that in place, we are now able to add the core remote config service
to the cluster-agent, as well as an instance of an "agent client" that
communicates directly to the core service instead of over a secure gRPC
connection. This serves as a starting point for any team wanting to build out RC
functionality in the cluster-agent.